Removed <c-typedef-type> from <icky-type-name>.
That <icky-type-name> is the result of the Lisp-style output
of the c-parser-engine. We should introduce a type hierarchy
and generate instances of that directly in the productions,
instead of passing this intermediate form.

Instances of pair(#"function", ...) in c-parser-engine.input
should be replaced by make(<c-function-declarator>, ...)
where <c-function-declarator> is one of the classes from
the type hierarchy replacing <icky-type-name>.

I suggest <c-declarator> as the new name of <icky-type-name>.