Initial check-in.
We're starting to factor out parts of the CMU optimizer.

The fer-transform library will eventually implement everything needed to
make FER safe for cback. The code in here will come out of the current
optimizer as we figure out what we need and how to isolate it from the
reoptimization loop.

This check-in contains work by myself and igor. xep.dylan is from the
original CMU optimizer library. For now, none of this code is currently
compiled by the regular build process.