Replaced the mostly broken common/file-system library with a largely working one:  I have implemented all of the file-system module as described in Chapter 8 of the Functional Objects, Inc. System and I/O document ... with the exception of 8.5 (home-directory, root-directories, etc), and the file- properties part of 8.6 (file-exists? and file-type work and are in use in this library).
To do this I moved the file-system module from d2c/compiler/base to
common/file-system and renamed it to base-file-system.

In order to follow Fun-O's <file-system-error> signature, I needed to add
<simple-condition> to d2c/runtime/dylan/condition.dylan (interposing it between
<format-string-error> and the other simple error types.

That was the ground work.  Then I gutted common/file-system and added the
implementation code, relying on lstat (from /usr/include/sys/stat.h) and
/usr/include/dirent.h's functions to do the primitive file and directory work.
I built the rest of the implementation in dylan to conform to the file-system
interface of common/file-system/exports.dylan.