Completed the Time-IO module and the Parse-Time Directives (instead of referring the user to K&R, I wrote them out SGML-hand, for a mind is a terrible thing to waste ... pondering C code).
Brought the Format library in sync with the rest of the manual (the
special "-- Function" tag in the index was too glaring for me), added the
format Directives into the documentation, because "We no need no STINKIN'
Manuals!" to look up these (for us Dylan Hackers) known-by-heart control
characters, and a Dylan newbie should have these directives instantly
accessable.  Also documented the %m directive ... pretty neat one, especially
if you want to print a time in a statement (a tested example is forthcoming).