Several minor changes.
The major minor change was that I replaced all local bindings to
'element' to be 'elemnt' instead, as [] (sequence element referencing)
uses the function 'element' (and, in some places in this file, a method
had both a local binding and an array reference -- *shudder*).

Corrected a misspelt word in a comment line.