Disable the special-casing of num.c if the compiler is egcs.
It's not clear to me exactly what the original fix was.  Is
it restricting the optimization level to -O2?  We're not
using higher than that right now *anyway*.  It would have
been better if the fix had built the currently standard
options line and then done a s/-O[3456789]/-O2/ on it, so
that it automatically tracks other changes in the
standard flags.