Did not encounter a <return> in control flow... error still appears on several of the phrases.
I am about to modularize the interpreter to have a macro that
implements the major nitty gritty parts but still leaves enough
tweakability to insert custom handling for several <region>s and
tracing functionality.
That is, we could simply put in a command-line debugger (or a html-based
one if you want) or a time machine that can also step backwards.

Also, it will feature an evaluation function that just updates bindings, and
returns results by non-local exit. (This was Andreas' idea, makes very much of sense)