Send d2c's *debug-output* (progress informational messages and debugging of d2c itself) to stdout instead of stderr.
Send warnings and errors about the program being compiled to
stderr instead of the *debug-output* (which used to be stderr
but now isn't).

This is more consistent with other compilers, and the build system
at my job assumes it, in supressing stdout when building an entire
source tree so that only diretory names and errors are shown.