Fixed bug where compiler crashes with "No element nn in 'xxxxxx'" in describe-source-location.  The fix is to put full paths to source files into library dump files, so that the original source code can be found when expanding macros and inline functions.
An alternative solution might be to include the source fragments
themselves in the .du file.

It might also be nice to give the user a command-line option to
change the prefix for where library source code can be found.

Because only file names were being stored in the .du file, none of the
library's source files could be found and this was usually silently
ignored.  But if the current directory happened to contain a file with
the same name then that was used, with bad results when trying to extract
text fragments.  This mostly happens with macros, so usually we crashed
if the library being compiled contained a macros.dylan.

Have removed temp code, including Andreas' exception catcher.