After a *lot* of wrestling, finally I get a "hello, world" window.
If you try to compile this, remember to unlimit the stack size, e.g.

$ ulimit -s unlimited

if you're using bash.

Two things don't work yet as they should. Number one: callbacks require
the usage of callback-method, which is inconvenient. Number two: return
values of functions returning a subclass of GObject get typed with the
formal type, not the actual type, so casts are needed where the compiler
could do things automatically.