Bug: 7016
Added a production to allow a type name token for the name of a
struct/union member in addition to an identifier. This is in line with
the ANSI C grammar found in The Annotated ANSI C Standard, and allows
a struct member to have the same name as a type -- they are
effectively in separate namespaces.