Job: website
Fixed two minor type-setting errors: "define generic" and "define
macro" weren't properly wrapped in CODE tags. This visual error occurs
in the printed DRM, so technically these might be considered errata
items, but it's a relatively minor change in appearance and fixes an
important inconsistency in the use of HTML tags.