Job: website
Corrected the location of an A element. It was outside a CODE element,
but it should be inside. This effects no significant difference to the
content but is important for correct interaction with the CSS
stylesheet (unfortunately; someday we should clean all this up and use
class/div/span to mark up everything appropriately.)