Job: website
Major changes to the markup of the index, to apply more semantic
markup and to make it look like the printed DRM. It is now a series of
nested lists (UL, but without bullets displayed), and CODE tags are
properly applied. In the previous version everything was marked
preformatted (PRE), whitespace was used to indent the hierarchy, and
everything rendered in monospaced code font.