job: 7094 Made more copy-sequence methods utilize check-start-end-bounds method.
Changed check-start-end-bounds to return only one parameter.

Implemented a shallow-copy for <sequence> which calls copy-sequence
if the sequence is not a range. Specializing on <mutable-sequence> would
have the same effect, at least for the pre-defined collection classes.
This method implementation is only an optimization: The default method
is slower but works. Maybe this method should be deleted later after map
is optimized more.
The DRM does not specify how shallow-copy should be implemented, so
implementing it by calling through to copy-sequence seems valid
(except for the issue about ranges).

For shallow copy, the DRM states "There is a predefined method
(for shallow-copy) for instances of <collection>... which creates a new
object by calling make on the type-for-copy of collection and filling
it with the same elements as collection".

Unfortunately that is not very detailed, and multi-dimensional arrays
are a wart, because they cannot be instantiated with a just a size: keyword.
Perhaps <array> should not inherit from <sequence>, or <sequence>
should be renamed <simple-sequence>.  I think there's a lot of issues with
the DRM's collections that have not been thought out as much
as they could be.