Job: website
Complete redesign of the online DRM, including a new navigation bar on
the left, which features convenient access to interior links to items
on each page (this was previously relegated to the bottom of the page,
and is particularly useful for navigating longer, more complex pages).

I also used CSS to make the printed rendition of the HTML more
reasonable. All the navigation items are hidden, and links are printed
in plain text, without underlines or blue coloring.

While I put CSS to good use, I also tried to make the HTML produce a
more reasonable, usable rendition than before, when CSS isn't
supported or stylesheets are disabled.

Overall, the layout and typography is now much closer to the printed
DRM, and I've reconstructed some content that was apparently lost
during the automatic translation to HTML.

I made use of CSS to hide some information not in the printed DRM or
that was just cluttering things up, while leaving it in the HTML in
case we want to make use of it in the future. (e.g., in the navigation
links, it includes tags like "[Open Generic Function]", which are no
longer displayed.)

I consolidated all the HTML pages of Appendix B, Exported Names, into
one page. There was no reason for it to be split up and it just made
it harder to navigate.

I added "disabled" renditions of each navigation button image, so the
set of buttons doesn't change on pages where one or more buttons don't
apply.

Too many fixes and adjustments to tagging to mention, but this
includes fixing the tagging of all the function signatures (though not
all the G.F. method signatures).

In fact, this work began as merely an attempt to fix the signature
tagging, but once you start making global changes to over a hundred
HTML files in dire need of cleanup, it's hard to find a good stopping
point. Or maybe that's just me.