Job: website
Fixed a number of links to chapters, which were pointing at the page
before the chapter start. Along the way, I also changed a number of
links that referred to named anchors to simply point at the file,
because the anchors are at the top of the file, anyway, and pointing
into the file usually cause browsers to scroll down to show the exact
line the link points to, which is undesirable when the user is
navigating to the start of a chapter.