Job: website
Fixed the layout of definition headings (which contain right-aligned "[Open Foo]"
attributes tags) in Firefox/Mozilla, which was displaying the attributes tag at the right,
but shifted down instead of aligned with the definition name to the left.

The original looked fine in Safari and Omniweb, and it isn't entirely clear to me which
behavior was correct according to the CSS spec, but this new CSS code produces the desired
effect in Safari and Firefox.

Note that G.F. Method signature "headings" still suffer from this problem in Firefox, but
fixing that will require adding tags to all the headers to identify the signature text so
we can use the same CSS approach with them.