Job: website
Added an "AddCharset" command to .htaccess so the HTTP server returns a character set
header for the DRM HTML files (set to "utf-8").

We probably need to decide on a policy for the entire website; do we tell Apache to
default to some particular value, e.g., ISO-8859-1 or UTF-8, and require files to either
use that character set or override it (with META tags, or .htaccess for whole directories)
when necessary?