Job: website
Added image maps for the class hierarchy figures so each of the class boxes are links to
the documentation for that class.

The new drm_Figure*.html files are generated by OmniGraffle from the .graffle files, then
their content is manually copied and adapted into the appropriate DRM page (OmniGraffle
doesn't generate XHTML -- it generates some older HTML version -- plus, the NAME attribute
of the maps are generic and need to be given appropriate names, so the output requires
some further editing.) If the OmniGraffle files are edited, they should be re-exported to
the various formats (PNG, PDF, HTML image maps) and the relevant portions of the image
maps integrated again.

This change includes a PDF and a PNG that were re-exported during my work, but they may
not actually have any visible changes from previous verions. I'm checking them in for
safety, in case they changed in a way that I can't immediately see.

Also added LANG attributes to the HEAD tags to indicate the content language (en-US).