Job: website
Backed out the rest of change 9916, which, while making it easier to test
the website locally--without putting it on the live server--unfortunately
broke the CGIs that also use these files but require absolute paths.

It would be really nice if we could figure out how to achieve easier local
testing in the future.