Job: website
Lots of changes to the online DRM.

- A completely new navigation area for the DRM. Unlike the previous version,
  this one contains a complete hierarchical contents listing.

- Many changes to better match the layout and appearance of the DRM, too
  numerous to list in detail. Notably, the typesetting of all the method
  signatures now matches the DRM (e.g., arguments and values are
  italicized).

- Some improvements in existing CSS rules to accomplish the same goals with
  better use of CSS.

- Some improvements in compatibility with various browsers.

Compatibility Note: Looks best in Safari and Firefox (and other browsers
that share the same engines). In Opera, the top of the navigation area is
placed over the book content instead of on the left, which is a severe
impediment, but as far as I can tell it's a bug in Opera and I see no easy
workaround. Likewise, it still looks bad in Windows IE 6 because IE doesn't
support 'position: fixed' and places the navigation area at the top of the
content. IE also doesn't support some CSS selectors, so the style sheet has
no effect on some portions of the page. These are all non-trivial to
address, unfortunately, but compatibility with these (frankly, flawed)
browsers is "an important future direction".