Job: website
Second-level bulleted list items now use a hollow square for the marker,
just like the printed DRM (instead of a hollow circle). This is
accomplished by replacing the marker with a Unicode character.

Formatted footnotes to match the printed DRM.

Renamed various "wm-div-..." classes to be clearer, changed them from
classes to ids, and moved the start of DIV id wm-div-body to after the
header DIV. Now the pages are uniformly constructed of three or four major
DIVs identified by id.