Job: website
Fixed the implementation of hollow square bullets and the indentation of
chapter tables of contents for Firefox (1.5).

Firefox doesn't support 'float: left' within a :before with
content:. Instead, it requires faking the float using a negative left
margin and approximating the correct spacing with margin-right:. The
'float: left' is still specified, so the layout is always correct on
browsers that support it -- the workaround can concievably end up placing
the start of the text slightly to the right or left of the left edge of the
remaining lines in the paragraph.

Firefox requires a more specific selector than Safari (2.0.2) in order for
some rules to override other rules. I don't know which is correct, but the
more verbose selector should always work correctly. This corrects the
indenting of second-level list items in the tables of contents at the start
of chapters.