Job: website
Removed anchor ids from links to the starts of files.

Most of the HTML files have an anchor named "HEADING@@@-0" (where "@@@" is
the number of the file) at the start of the content, and the
(machine-generated) links to those files are of the form
"filename#HEADING@@@-0". This adds unnecessary noise to the URL and can
result in browsers displaying the destination page scrolled down a little
bit. Now, those links simply point to the file.