Job: website
Removed the redundant class T0.Text0. It was used to mark paragraphs that
followed a header, which can be determined without the aid of a
class. (There was also one erroneous use of it, which caused two paragraphs
to run together in the printed DRM, and a couple of uses where it had no
effect.)