Job: website
In the Errata page, quoted text from the DRM was wrapped in list
items. I've replaced the one-item lists with BLOCKQUOTE, which is
semantically more correct and lays out better in browsers.

I also made some minor changes to the errata text to make it clearer and to
use consistent wording.