job: fd
Cleanup of documentation. It used to consist of a number of tarballs with different manuals. The contents of the tarballs overlaps, so I extracted the manuals from each one and removed the tarballs. This eliminates duplicates, and makes the HTML editable. Also, compiled HTMLHelp files have been dropped: again they are overlapping, and can be regenerated from the HTML files. Finally, a new master index file for Open Dylan has been added.