Bug: 7355
Dylan Mode now indents multiple local methods correctly when "local method" is
on one line. It indents all the methods to align with the first occurrence of
the word "method".

The primary problem was that, while searching backwards for the start of the
current expression/constituent, it stopped at "end," because the regular
expression erroneously required at least one space ("end[ \t]+"). I changed it
to "end[ \t]*" and now it can search backwards to locate the start of the
"local" constituent.

With that resolved, I added some code that gets the column position of the "m"
in "local method" and returns that as the indentation depth when the "local"
constituent spans multiple lines.