Job: website
In the navigation area, the current item is now a link to the current
page instead of just text. This is more symmetrical (everything in the
navigation area is a link) and provides a convenient way to go to the
top of long pages.

(They were originally links, long ago, but I removed the links,
thinking you didn't need links to the current page, and it made them
stand out visually from the links. Since then, I've changed my mind,
and I've learned how to display them in black instead of the link
color via CSS.)