Job: fd
support for more commands:
 *quit-lisp - that was easy
 *default-directory/set-default-directory: sets working directory
 *list-all-package-names: for completion... here we list all library
   names
 *set-package: sets package to given string (and opens this package)

The problem we now face is that LISP only has packages, where Dylan has
libraries and modules... thus, a package has probably to be a library and
a module (library:module as in the Dylan IDE). This information is sadly
not available; we have two different cases:
  a .dylan file is opened -> there we have the "module:" header, and we
    want to find definitions/descriptions of various forms
  set-package or in-package is called and something in the lisp shell
    needs to be described. here we have a library name from the completion
    of the defined registries, but we have no modules.

If we open all compiler databases and ask the compiler for all modules of
a given library/project, this takes really long. And a startup of > 2s of
dswank is not usable anymore.

In the former case we currently have the assumption that the project in
development is open. This can either be done by ", set-package" or via
"open" in the slime shell.

The natural way would be to have separate commands for setting library
and module (slime would have to be changed...). Or some magic voodoo
in dylan-mode which provides the knowledge what library the file belongs
to.