Job: minor
QV directives have been simplified. They no longer automatically apply the API style to API links. Corresponding changes to templates.
Special-cased QV directives that use the quoted text as a target where the target is an init-keyword, argument, or value. They do not replace the quoted text with the section title.
Fixed a bug where a QV directive that uses the quoted text as a target causes an error when target is not found. Correct behavior is to just give a warning. Moved this logic out of QV markup translation into XREF resolution.