Job: minor
General fixes:
* Changed declaration of <case-insensitive-skip-list> for r12813.
* Simplified api-not-found-in-code warning and removed unnecessary upper-casing code.
* Removed spurious unused-docs-in-topic and section-id-without-topic-id warnings.
* Account for possibly invalid user-supplied fully qualified names.

After experimentation, I realize that catalog topics need both [LIST OF] directives and a normal list of child topics. The directive needs to give synonyms and aliases, and the child list ought to be present to navigate to APIs added under various catalog pages. Plus, this means I don't need trickery to generate DITA output similar to the HTML output.

Fixes for that:
* Removed exclusion of API topics from topics' child topic list.
* Added namespace-specific titles to API topics.
* Allow <xref> to enclose <title-seq> directly, instead of only title <conref>.