Job: gd
Allow mindy to use a LOT of memory (200M) in between runs of the GC. Prior to this change, GC was taking approximately 50% of the CPU time while doing builds. This reduces the GC time significantly and vastly speeds up the build. It may be an issue though for someone trying to bootstrap on an embedded platform.