Job: gd
On Mac OS X, dlfcn uses bool which is defined in stdbool.h. stdbool.h checks against __STDC_VERSION__ and __GNUC__ to see if it needs to define _Bool. We define __STDC_VERSION__ to a value that should have _Bool though, so stdbool.h does not define it. We just define it here instead.