Bug: 7495
Apply the changes in commits #13009, #13011, and #13012 to d2c 2.5.

The 2.5 branch seems to have a much higher variability in build
times -- I'm seeing up to 20 - 30 seconds difference on sucessive
runs.

The best before run was 2m25.9, and after 1m47.8, an improvement
of 38.1 seconds or 35.3% faster. This is considerably less in
both absolute and relative terms than the improvement in d2c 2.4.